+++ src/style/myPage.scss
@@ -7,6 +7,9 @@
   
 }
 
+.off {
+  display: none;
+}
 
 .all_content{
   position: relative;
@@ -20,9 +23,6 @@
     margin-right: 15px;
     // vertical-align: middle;
     }
-  .off {
-    display: none;
-  }
   .top_box{
     left: 100px;
     // width: 1200px;
@@ -157,7 +157,7 @@
               text-align: center;
               font-size: 14px;
               font-weight: 400;
-              text-align: left;
+              // text-align: left;
               thead {
                 th {
                   color: #98989f;
